refactor(engine): ArtistStateProcessor — weekly artist/executive state cluster
Phase 2 engine-seams PR-11 (Commit A of 2). Verbatim move of the artist-state
domain out of GameEngine into shared/engine/processors/ArtistStateProcessor.ts:
applyArtistChangesToDatabase, processWeeklyMoodChanges (+ calculateReleaseMoodBoost,
calculateWorkloadStressPenalty, calculateNaturalMoodDrift), processWeeklyPopularityChanges,
processExecutiveMoodDecay (+ formatExecutiveRole), selectHighestPopularityArtist,
loadSignedArtists.

Bodies are character-identical; only this. rebound to WeekContext. The lone
seeded draw (selectHighestPopularityArtist tie-break) moves from raw this.rng()
to ctx.getRandom(0,1) — the same draw from the same seeded stream in the same
pipeline position (getRandom(0,1) === rng()).

Engine keeps same-signature private delegates for every method so the advanceWeek
pipeline and the mood test suites (which call these via (engine as any).method)
keep working byte-for-byte. The three mood helpers and formatExecutiveRole had no
callers outside their now-moved parents, so they are deleted from the engine.

processDelayedEffects is deferred to Commit B (it calls applyEffects, the hub,
which lands with ActionProcessor) to keep each commit independently green.

game-engine.ts 2667 -> 2238. Golden master 8/8 (zero updates, x2); mood + engine
suites green (12 files / 158 tests).
